feature HWD-014 Code generator based kotlin metadata.
Having an hybrid generator was proving to be highly complex and it was easier to split the implementations.
Now there is a kotlin metadata based generator that should contain full support for kotlin types and
a plain java generator that will act as a fallback for when kotlin metadata is not available but has serious limitations.